+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### master
+* Update dependencies
+* Update `renderer` & `inline-compilers` for `sham-ui^4.0.0` 
+
 ### 3.3.0
 * Merge `compileWith` to `compile` & `compileAsSFCWith` to `compileAsSFC`